Contribution template (#84)
* Add Template file to wiki root and update README.md

This update provides a sample markdown file for contributing to the 
wiki. It also updates the README to make contributors aware of the 
template's existence.

The template also contain links to some references for working with 
markdown and MkDocs admonition support.

* contribution-template.md reference
Added reference in CONTRIBUTING.md for the contributing-template.md 
file.