Add page on why one might use DNS related features
Rather than a detailed guide on how to implement any particular feature,
this guide talks about how various DNS features can be implemented and
what you can expect from them. This page looks from the simple (mDNS)
to the complex (SSHFP) and why they are (or are not) useful.

The purpose of this guide is to help self hosters determine what level
of complexity makes sense in their environement.